+++ packages/ingest/sftp.js
@@ -6,6 +6,8 @@ const { log } = require('@cumulus/common');
 const Crypto = require('./crypto').DefaultProvider;
 const recursion = require('./recursion');
 const { omit } = require('lodash');
+const get = require('lodash.get');
+const { S3, KMS } = require('./aws');
 
 module.exports.sftpMixin = (superclass) => class extends superclass {
 
@@ -17,7 +19,9 @@ module.exports.sftpMixin = (superclass) => class extends superclass {
       host: this.host,
       port: this.port || 22,
       user: this.username,
-      password: this.password
+      password: this.password,
+      privateKey: get(this.provider, 'privateKey', null),
+      cmKeyId: get(this.provider, 'cmKeyId', null)
     };
 
     this.client = null;
@@ -34,6 +38,23 @@ module.exports.sftpMixin = (superclass) => class extends superclass {
       if (this.username) {
         this.options.user = await Crypto.decrypt(this.username);
         this.decrypted = true;
+      }
+    }
+
+    if (this.options.privateKey) {
+      const bucket = process.env.internal;
+      const stackName = process.env.stackName;
+      // we are assuming that the specified private key is in the S3 crypto directory
+      log.debug(`Reading Key: ${this.options.privateKey} bucket:${bucket},stack:${stackName}`);
+      const priv = await S3.get(bucket, `${stackName}/crypto/${this.options.privateKey}`);
+
+      if (this.options.cmKeyId) {
+        // we are using AWS KMS and the privateKey is encrypted
+        this.options.privateKey = await KMS.decrypt(priv.Body.toString());
+      }
+      else {
+        // private key is not encrypted...
+        this.options.privateKey = priv.Body.toString();
       }
     }